refactor to hashIndexBuilder and hashIndexReader
